@@ -32,8 +32,6 @@ var podNamespace = os.Getenv("POD_NAMESPACE")
 var isMaster bool                  // indicates this instance is the master and should be running checks
 var upcomingMasterState bool       // the upcoming master state on next interval
 var lastMasterChangeTime time.Time // indicates the last time a master change was seen
-var listenNamespace string         // namespace to listen (watch/get) `khcheck` resources on.  If blank, all namespaces will be monitored.
-
 // Interval for how often check pods should get reaped. Default is 30s.
 var checkReaperRunInterval = os.Getenv("CHECK_REAPER_RUN_INTERVAL")
 
@@ -64,8 +62,8 @@ var khCheckClient *khcheckv1.KHCheckV1Client
 var khJobClient *khjobv1.KHJobV1Client
 
 // constants for using the kuberhealthy status CRD
-const stateCRDGroup = "comcast.github.io"
-const stateCRDVersion = "v1"
+// const stateCRDGroup = "comcast.github.io"
+// const stateCRDVersion = "v1"
 const stateCRDResource = "khstates"
 
 // constants for using the kuberhealthy check CRD
@@ -73,96 +71,11 @@ const checkCRDGroup = "comcast.github.io"
 const checkCRDVersion = "v1"
 const checkCRDResource = "khchecks"
 
-// the global kubernetes client
+// kubernetesClient is the global kubernetes client
 var kubernetesClient *kubernetes.Clientset
 
-// Set dynamicClient that represents the client used to watch and list unstructured khchecks
+// dynamicClient represents the client used to watch and list unstructured khchecks
 var dynamicClient dynamic.Interface
-
-// setUpConfig loads and sets default Kuberhealthy configurations
-// Everytime kuberhealthy sees a configuration change, configurations should reload and reset
-func setUpConfig() error {
-	cfg = &Config{
-		kubeConfigFile: filepath.Join(os.Getenv("HOME"), ".kube", "config"),
-		LogLevel:       "info",
-	}
-
-	// attempt to load config file from disk
-	err := cfg.Load(configPath)
-	if err != nil {
-		log.Println("WARNING: Failed to read configuration file from disk:", err)
-	}
-
-	// set env variables into config if specified. otherwise set external check URL to default
-	externalCheckURL, err := getEnvVar(KHExternalReportingURL)
-	if err != nil {
-		if len(podNamespace) == 0 {
-			return errors.New("KH_EXTERNAL_REPORTING_URL environment variable not set and POD_NAMESPACE environment variable was blank.  Could not determine Kuberhealthy callback URL.")
-		}
-		log.Infoln("KH_EXTERNAL_REPORTING_URL environment variable not set, using default value")
-		externalCheckURL = "http://kuberhealthy." + podNamespace + ".svc.cluster.local/externalCheckStatus"
-	}
-	cfg.ExternalCheckReportingURL = externalCheckURL
-	log.Infoln("External check reporting URL set to:", cfg.ExternalCheckReportingURL)
-	return nil
-}
-
-// setUp loads, parses, and sets various Kuberhealthy configurations -- from flags, config values and env vars.
-func setUp() error {
-
-	var useDebugMode bool
-
-	// setup flaggy
-	flaggy.SetDescription("Kuberhealthy is an in-cluster synthetic health checker for Kubernetes.")
-	flaggy.String(&configPath, "c", "config", "(optional) absolute path to the kuberhealthy config file")
-	flaggy.Bool(&useDebugMode, "d", "debug", "Set to true to enable debug.")
-	flaggy.Parse()
-
-	err := setUpConfig()
-	if err != nil {
-		return err
-	}
-
-	// parse and set logging level
-	parsedLogLevel, err := log.ParseLevel(cfg.LogLevel)
-	if err != nil {
-		err := fmt.Errorf("unable to parse log-level flag: %s", err)
-		return err
-	}
-
-	// log to stdout and set the level to info by default
-	log.SetOutput(os.Stdout)
-	log.SetLevel(parsedLogLevel)
-	log.Infoln("Startup Arguments:", os.Args)
-
-	// no matter what if user has specified debug leveling, use debug leveling
-	if useDebugMode {
-		log.Infoln("Setting debug output on because user specified flag")
-		log.SetLevel(log.DebugLevel)
-	}
-
-	// Handle force master mode
-	if cfg.EnableForceMaster == true {
-		log.Infoln("Enabling forced master mode")
-		masterCalculation.DebugAlwaysMasterOn()
-	}
-
-	// determine the name of this pod from the POD_NAME environment variable
-	podHostname, err = getEnvVar("POD_NAME")
-	if err != nil {
-		err := fmt.Errorf("failed to determine my hostname: %s", err)
-		return err
-	}
-
-	// setup all clients
-	err = initKubernetesClients()
-	if err != nil {
-		err := fmt.Errorf("failed to bootstrap kubernetes clients: %s", err)
-		return err
-	}
-
-	return nil
-}
 
 func main() {
 
@@ -173,8 +86,7 @@ func main() {
 	}
 
 	// Create a new Kuberhealthy struct
-	kuberhealthy := NewKuberhealthy()
-	kuberhealthy.ListenAddr = cfg.ListenAddress
+	kuberhealthy := NewKuberhealthy(cfg)
 
 	// create run context and start listening for shutdown interrupts
 	khRunCtx, khRunCtxCancelFunc := context.WithCancel(context.Background())
@@ -196,7 +108,7 @@ func listenForInterrupts(k *Kuberhealthy) {
 	sigChan := make(chan os.Signal, 1)
 
 	// register for shutdown events on sigChan
-	signal.Notify(sigChan, os.Interrupt, syscall.SIGTERM, syscall.SIGKILL)
+	signal.Notify(sigChan, os.Interrupt, syscall.SIGTERM)
 	log.Infoln("shutdown: waiting for sigChan notification...")
 	<-sigChan
 	log.Infoln("shutdown: Shutting down due to sigChan signal...")
@@ -254,7 +166,7 @@ func initKubernetesClients() error {
 	khJobClient = jobClient
 
 	// make a dynamicClient for kubernetes unstructured checks
-	restConfig, err := clientcmd.BuildConfigFromFlags("", configPath)
+	restConfig, err := clientcmd.BuildConfigFromFlags(kc.RESTClient().Get().URL().Host, configPath)
 	if err != nil {
 		log.Fatalln("Failed to build kubernetes configuration from configuration flags:", err)
 	}
@@ -262,6 +174,96 @@ func initKubernetesClients() error {
 	dynamicClient, err = dynamic.NewForConfig(restConfig)
 	if err != nil {
 		log.Fatalln("Failed to create kubernetes dynamic client configuration")
+	}
+
+	return nil
+}
+
+// setUpConfig loads and sets default Kuberhealthy configurations
+// Everytime kuberhealthy sees a configuration change, configurations should reload and reset
+func setUpConfig() error {
+	cfg = &Config{
+		kubeConfigFile: filepath.Join(os.Getenv("HOME"), ".kube", "config"),
+		LogLevel:       "info",
+	}
+
+	// attempt to load config file from disk
+	err := cfg.Load(configPath)
+	if err != nil {
+		log.Println("WARNING: Failed to read configuration file from disk:", err)
+	}
+
+	// set env variables into config if specified. otherwise set external check URL to default
+	externalCheckURL, err := getEnvVar(KHExternalReportingURL)
+	if err != nil {
+		if len(podNamespace) == 0 {
+			return errors.New("env KH_EXTERNAL_REPORTING_URL not set and POD_NAMESPACE environment variable was blank")
+		}
+		log.Infoln("KH_EXTERNAL_REPORTING_URL environment variable not set, using default value")
+		externalCheckURL = "http://kuberhealthy." + podNamespace + ".svc.cluster.local/externalCheckStatus"
+	}
+	cfg.ExternalCheckReportingURL = externalCheckURL
+	log.Infoln("External check reporting URL set to:", cfg.ExternalCheckReportingURL)
+	return nil
+}
+
+// setUp loads, parses, and sets various Kuberhealthy configurations -- from flags, config values and env vars.
+func setUp() error {
+
+	var useDebugMode bool
+
+	// setup global config struct
+	err := setUpConfig()
+	if err != nil {
+		return err
+	}
+
+	// setup flaggy
+	flaggy.SetDescription("Kuberhealthy is an in-cluster synthetic health checker for Kubernetes.")
+	flaggy.String(&configPath, "c", "config", "Absolute path to the kuberhealthy config file")
+	flaggy.Bool(&useDebugMode, "d", "debug", "Set to true to enable debug.")
+	flaggy.Bool(&cfg.EnableForceMaster, "", "forceMaster", "Set to force master responsibilities on.")
+	flaggy.Parse()
+
+	// parse and set logging level
+	parsedLogLevel, err := log.ParseLevel(cfg.LogLevel)
+	if err != nil {
+		err := fmt.Errorf("unable to parse log-level flag: %s", err)
+		return err
+	}
+
+	// log to stdout and set the level to info by default
+	log.SetOutput(os.Stdout)
+	log.SetLevel(parsedLogLevel)
+	log.Infoln("Startup Arguments:", os.Args)
+
+	// no matter what if user has specified debug leveling, use debug leveling
+	if useDebugMode {
+		log.Infoln("Setting debug output on because user specified flag")
+		log.SetLevel(log.DebugLevel)
+	}
+
+	// Handle force master mode
+	if cfg.EnableForceMaster {
+		log.Infoln("Enabling forced master mode")
+		masterCalculation.DebugAlwaysMasterOn()
+	}
+
+	// determine the name of this pod from the POD_NAME environment variable
+	podHostname, err = getEnvVar("POD_NAME")
+	if err != nil {
+		err := fmt.Errorf("failed to determine my hostname: %s", err)
+		return err
+	}
+
+	// determine the name of this pod from the POD_NAME environment variable
+	cfg.TargetNamespace = os.Getenv("TARGET_NAMESPACE")
+
+	// setup all clients
+	err = initKubernetesClients()
+	if err != nil {
+		err := fmt.Errorf("failed to bootstrap kubernetes clients: %s", err)
+		return err
 	}
 
 	return nil
